+++ src/renderer/shared/wallet/ETH/EthWallet.js
