+++ src/renderer/shared/wallet/NEO/NeoMnemonic.js
