+++ src/renderer/shared/wallet/WalletHelpers.js
