+++ src/renderer/shared/wallet/common/BIP44.js
@@ -1,6 +1,22 @@
-import { DEFAULT_COIN } from 'shared/values/coins';
+import { includes } from 'lodash';
+import bip32 from 'bip32';
 
-// TODO merge with /wallets/common/BIP44
+import { DEFAULT_COIN, COIN_TYPES } from 'shared/values/coins';
+
+const deriveChild = ({ wallet, seed }) => {
+  const { coinType, index, account, change } = wallet;
+  if (!includes(COIN_TYPES, coinType)) throw new Error('No valid coin type was given.');
+
+  const root = bip32.fromSeed(seed);
+
+  return root
+    .deriveHardened(44) // Purpose (bip44)
+    .deriveHardened(coinType) // Coin type https://github.com/satoshilabs/slips/blob/master/slip-0044.md
+    .deriveHardened(account) // Account (different account levels)
+    .derive(change) // Change (0 = external/public view, 1 = internal chain/private view)
+    .derive(index); // Address index
+};
+
 function assertPositiveInteger(input, inputName) {
   if (!Number.isInteger(input) || input < 0) {
     throw new Error(`${input} is an invalid input for ${inputName}`);
@@ -15,7 +31,7 @@ function to8BitHex(num) {
 /**
  * Returns a BIP44 hex string.
  */
-export function BIP44({ address = 0, change = 0, account = 0, coinType = DEFAULT_COIN }) {
+function BIP44({ address = 0, change = 0, account = 0, coinType = DEFAULT_COIN }) {
   assertPositiveInteger(address, 'address');
   assertPositiveInteger(change, 'change');
   assertPositiveInteger(account, 'account');
@@ -34,4 +50,5 @@ export function BIP44({ address = 0, change = 0, account = 0, coinType = DEFAULT
   ].join('');
 }
 
-export default BIP44;
+// eslint-disable-next-line
+export { deriveChild, assertPositiveInteger, to8BitHex, BIP44 };
