+++ src/renderer/shared/wallet/common/Ledger.js
@@ -1,6 +1,6 @@
 import LedgerNode from '@ledgerhq/hw-transport-node-hid';
 
-import { evalTransportError, BIP44, VALID_STATUS, assembleSignature } from './Wallet/common/LedgerHelpers';
+import { evalTransportError, BIP44, VALID_STATUS, assembleSignature } from 'shared/wallet/common/LedgerHelpers';
 
 export default class Ledger {
   constructor(path) {
