+++ src/renderer/shared/wallet/common/index.js
