+++ src/renderer/shared/wallet/index.js
@@ -1,4 +1,3 @@
 import Wallet from './Wallet';
 
-// TOOD move whole wallet section to shared/wallet
 export default Wallet;
